+++ src/Entity/Initiative.php
@@ -71,6 +71,16 @@ class Initiative extends AbstractEntity
     #[ORM\JoinTable(name: 'initiative_contact')]
     private Collection $contacts;
 
+    /**
+     * Not cascade-validated, as on the other free-tagging collections: a violation
+     * would carry the path partners[0].name, which the single text input cannot render.
+     *
+     * @var Collection<int, Partner>
+     */
+    #[ORM\ManyToMany(targetEntity: Partner::class, cascade: ['persist'])]
+    #[ORM\JoinTable(name: 'initiative_partner')]
+    private Collection $partners;
+
     /** @var Collection<int, InitiativeImage> */
     #[ORM\OneToMany(targetEntity: InitiativeImage::class, mappedBy: 'initiative', cascade: ['persist', 'remove'], orphanRemoval: true)]
     private Collection $images;
@@ -116,6 +126,7 @@ class Initiative extends AbstractEntity
         parent::__construct();
         $this->strategies = new ArrayCollection();
         $this->contacts = new ArrayCollection();
+        $this->partners = new ArrayCollection();
         $this->stakeholders = new ArrayCollection();
         $this->tags = new ArrayCollection();
         $this->images = new ArrayCollection();
@@ -281,6 +292,28 @@ class Initiative extends AbstractEntity
     public function removeContact(Contact $contact): static
     {
         $this->contacts->removeElement($contact);
+
+        return $this;
+    }
+
+    /** @return Collection<int, Partner> */
+    public function getPartners(): Collection
+    {
+        return $this->partners;
+    }
+
+    public function addPartner(Partner $partner): static
+    {
+        if (!$this->partners->contains($partner)) {
+            $this->partners->add($partner);
+        }
+
+        return $this;
+    }
+
+    public function removePartner(Partner $partner): static
+    {
+        $this->partners->removeElement($partner);
 
         return $this;
     }